requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tokenizers>=0.13.3
torch-geometric>=2.3.0
pgmpy>=0.2.10
librosa>=0.9.2
SpeechRecognition>=3.8.1
opencv-python>=4.7.0.72
neo4j>=5.10.0
rdflib>=6.3.2
stable-baselines3>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
Flask
numpy
librosa
pydub
tqdm
werkzeug
keras-facenet
googletrans
openai-whisper
ffmpeg
